# 07_api_design/openapi.yaml
openapi: 3.0.3
info:
  title: Atelier API
  version: 1.0.0
  description: |
    Atelier - AI 社員常駐型開発プロジェクト統合管理 SaaS の API。

    ★ API 契約凍結: 2026-05-20T13:30:00+09:00 ★
    以後の変更は ADR + 影響解析 + 承認フロー必須。

    認証 4 系統:
    - 一般 JWT (Supabase Auth)
    - クライアント別 JWT (role=client_portal, project_id クレーム限定・R-T08 対応)
    - MCP Token
    - Service Role (バックエンド内部 + Bridge worker)
  contact:
    name: ENGINE BASE
    email: info@engine-base.com
  license:
    name: Proprietary
servers:
  - url: https://atelier.app/api/v1
    description: 本番
  - url: https://*.preview.atelier.app/api/v1
    description: Preview Deploy
  - url: http://localhost:8000/api/v1
    description: ローカル開発
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bridgeToken:
      type: apiKey
      in: header
      name: X-Bridge-Token
    clientPortalAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - UNAUTHENTICATED
                - INVALID_CREDENTIALS
                - INVALID_BRIDGE_TOKEN
                - INVALID_INVITATION_TOKEN
                - FORBIDDEN
                - ACCOUNT_LOCKED
                - ACCOUNT_DELETED
                - NOT_FOUND
                - CONFLICT
                - INVALID_LIFECYCLE_STAGE
                - DEPENDENCIES_NOT_MET
                - ALREADY_DECIDED
                - EMAIL_ALREADY_REGISTERED
                - EMAIL_IN_DELETION_GRACE
                - OWNER_OF_WORKSPACE_WITH_OTHERS
                - INVITATION_EXPIRED
                - VALIDATION_ERROR
                - CONSENT_REQUIRED
                - RATE_LIMITED
                - BYOK_QUOTA_EXCEEDED
                - INTERNAL_ERROR
                - BRIDGE_OFFLINE
                - PARALLEL_LIMIT_REACHED
                - F_CTX01_TIMEOUT
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  code: { type: string }
                  message: { type: string }
            request_id: { type: string, format: uuid }
            trace_id: { type: string }
...
    Workspace:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        member_count: { type: integer }
        project_count: { type: integer }
        plan: { type: string, enum: [free, pro, enterprise] }
        deleted_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    InviteMemberRequest:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [owner, member, viewer] }
        message: { type: string, maxLength: 500 }

    # ========================================================================
    # project
    # ========================================================================
...
    WorkspaceMember:
      type: object
      properties:
        workspace_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string, nullable: true }
        role: { type: string, enum: [owner, member, viewer] }
        joined_at: { type: string, format: date-time }
...
    Phase:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        order: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        status: { type: string, enum: [pending, in_progress, completed, skipped] }
        started_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...
paths:
...
  /workspaces:
    get:
      tags: [workspaces]
      summary: 所属 WS 一覧
      x-implementation-path: apps/api/src/routes/workspaces.py::list
      x-screen-ids: [S-B01]
      responses:
        '200':
          description: WS 一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Workspace' }
    post:
      tags: [workspaces]
      summary: 新規 WS 作成
      x-implementation-path: apps/api/src/routes/workspaces.py::create
      x-screen-ids: [S-A03]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
                description: { type: string }
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Workspace' }
  /workspaces/{workspace_id}:
    get:
      tags: [workspaces]
      summary: WS 詳細
      x-implementation-path: apps/api/src/routes/workspaces/__init__.py::get_workspace
      x-screen-ids: [S-B01]
      parameters:
        - { name: workspace_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: WS 詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Workspace' }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      tags: [workspaces]
      summary: WS 更新
      x-implementation-path: apps/api/src/routes/workspaces/__init__.py::update_workspace
      x-screen-ids: [S-A03]
      parameters:
        - { name: workspace_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 2, maxLength: 50 }
                description: { type: string }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Workspace' }
        '403': { description: 権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [workspaces]
      summary: WS 削除（論理）
      x-implementation-path: apps/api/src/routes/workspaces/__init__.py::delete_workspace
      x-screen-ids: [S-A03]
      parameters:
        - { name: workspace_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: 削除成功 }
        '403': { description: 権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ==========================================================================
  # workspace-members (招待・ロール・削除)
  # ==========================================================================
  /workspaces/{workspace_id}/members:
    parameters:
      - { name: workspace_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [workspace-members]
      summary: WS メンバー一覧
      x-implementation-path: apps/api/src/routes/workspace_members/__init__.py::list_members
      x-screen-ids: [S-A03]
      responses:
        '200':
          description: メンバー一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/WorkspaceMember' } }
    post:
      tags: [workspace-members]
      summary: WS メンバー招待
      x-implementation-path: apps/api/src/routes/workspace_members/__init__.py::invite_member
      x-screen-ids: [S-A03]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InviteMemberRequest' }
      responses:
        '201':
          description: 招待成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/WorkspaceMember' }
        '403': { description: owner でない, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: 既にメンバー, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: メール未登録, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /workspaces/{workspace_id}/members/{user_id}:
    parameters:
      - { name: workspace_id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [workspace-members]
      summary: WS メンバーのロール変更
      x-implementation-path: apps/api/src/routes/workspace_members/__init__.py::update_role
      x-screen-ids: [S-A03]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [owner, member, viewer] }
      responses:
        '200':
          description: 変更成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/WorkspaceMember' }
        '403': { description: 不在 or 権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [workspace-members]
      summary: WS メンバー削除
      x-implementation-path: apps/api/src/routes/workspace_members/__init__.py::remove_member
      x-screen-ids: [S-A03]
      responses:
        '204': { description: 削除成功 }
        '403': { description: 不在 or 権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ==========================================================================
  # projects (8 endpoints) - 主要 3 つ
  # ==========================================================================
...
  /workflow/phases:
    get:
      tags: [workflow]
      summary: 工程一覧
      x-implementation-path: apps/api/src/routes/workflow/__init__.py::list_phases
      x-screen-ids: [S-G01]
      parameters:
        - { name: project_id, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Phase' } }
    post:
      tags: [workflow]
      summary: 工程作成
      x-implementation-path: apps/api/src/routes/workflow/__init__.py::create_phase
      x-screen-ids: [S-G01]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, order, name]
              properties:
                project_id: { type: string, format: uuid }
                order: { type: integer, minimum: 0 }
                name: { type: string, maxLength: 200 }
                description: { type: string }
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Phase' }
  /workflow/phases/{phase_id}:
    parameters:
      - { name: phase_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [workflow]
      summary: 工程詳細
      x-implementation-path: apps/api/src/routes/workflow/__init__.py::get_phase
      x-screen-ids: [S-G01]
      responses:
        '200':
          description: 詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Phase' }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      tags: [workflow]
      summary: 工程遷移・更新
      x-implementation-path: apps/api/src/routes/workflow/__init__.py::update_phase
      x-screen-ids: [S-G01]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 200 }
                description: { type: string }
                status: { type: string, enum: [pending, in_progress, completed, skipped] }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Phase' }
        '403': { description: 権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [workflow]
      summary: 工程削除
      x-implementation-path: apps/api/src/routes/workflow/__init__.py::delete_phase
      x-screen-ids: [S-G01]
      responses:
        '204': { description: 削除成功 }
        '403': { description: 権限なし, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 不在 or 不可視, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...